Move include up a level, and update include guards
Collapses common/src/ to just common/ since there's no longer another
directory in common/.

Also moves linux embedder.h into include/linux to unify include dir.
