+++ example/linux/flutter_embedder_example.cc
@@ -21,7 +21,7 @@
 
 #include <color_panel/color_panel_plugin.h>
 #include <file_chooser/file_chooser_plugin.h>
-#include <flutter_desktop_embedding/embedder.h>
+#include <flutter_desktop_embedding/linux/embedder.h>
 #include <menubar/menubar_plugin.h>
 
 namespace {
