+++ library/common/method_call.cc
@@ -1,0 +1,23 @@
+// Copyright 2018 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+#include "library/include/flutter_desktop_embedding/method_call.h"
+
+namespace flutter_desktop_embedding {
+
+MethodCall::MethodCall(const std::string &method_name)
+    : method_name_(method_name) {}
+
+MethodCall::~MethodCall() {}
+
+}  // namespace flutter_desktop_embedding
