+++ library/windows/GLFW Library.vcxproj.filters
@@ -18,7 +18,7 @@
     <ClCompile Include="embedder.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="..\shared\src\internal\plugin_handler.cc">
+    <ClCompile Include="..\common\internal\plugin_handler.cc">
       <Filter>Source Files</Filter>
     </ClCompile>
   </ItemGroup>
